MB-9 switch to SqLite database

+++ src/backend/DomainLogic/ValueObjects/UserID.cs
@@ -2,7 +2,7 @@
 {
     public sealed class UserID
     {
-        public decimal Value { get; private set; }
+        public int Value { get; private set; }
 
         public UserID(int value)
         {
